Use struct and fields to load the groups.
The code to load the groups is run on booting time so it should be fine
(performance-wise)
The type of each field still need to be added as a tag to most structs.
The description is still missing. I think It can also be added as a tag
but I'm not sure at the moment since it can make the code harder to
read.

Also, for now it is loading all groups but it'd be nice to make that
configurable.